requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
pytz